KTO-1007: Rename Lomake to HakukohdeKortti and move LocalizedHTML to separate file
